Rename com.example package to dev.kocken
