+++ runelite-plugin.properties
@@ -3,4 +3,4 @@ author=Nobody
 support=
 description=An example greeter plugin
 tags=
-plugins=com.example.ExamplePlugin
+plugins=dev.kocken.ExamplePlugin
